chore(annotations): Separate annotation processing into another module
