+++ annotations/src/main/java/dev/hypera/chameleon/annotations/Plugin.java
@@ -20,15 +20,50 @@
  *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  *  SOFTWARE.
  */
-package dev.hypera.chameleon.core.annotations;
+package dev.hypera.chameleon.annotations;
 
+import dev.hypera.chameleon.annotations.processing.generation.Generator;
+import dev.hypera.chameleon.annotations.processing.generation.impl.velocity.VelocityGenerator;
+import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import org.jetbrains.annotations.NotNull;
 
-@Retention(RetentionPolicy.RUNTIME)
-public @interface PlatformDependency {
+@Retention(RetentionPolicy.SOURCE)
+@Target(ElementType.TYPE)
+public @interface Plugin {
 
-    String name();
-    boolean soft();
+	@NotNull String id();
+	@NotNull String name() default "";
+	@NotNull String version();
+	@NotNull String description() default "";
+	@NotNull String url() default "";
+	@NotNull String[] authors() default {};
+	@NotNull PlatformDependency[] dependencies() default {};
+
+	@NotNull String logPrefix() default "[%s]";
+	@NotNull Platform[] platforms();
+
+
+	enum Platform {
+
+		BUNGEECORD(null),
+		MINESTOM(null),
+		SPIGOT(null),
+		VELOCITY(VelocityGenerator.class);
+
+		private final @NotNull Class<? extends Generator> generator;
+
+		Platform(@NotNull Class<? extends Generator> generator) {
+			this.generator = generator;
+		}
+
+		public @NotNull Class<? extends Generator> getGenerator() {
+			return generator;
+		}
+
+	}
+
 
 }
