+++ annotations/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,0 +1,1 @@
+dev.hypera.chameleon.annotations.processing.ChameleonAnnotationProcessor
